Distribute name change reviews over 24h
