+++ server/src/api/jobs/job.types.ts
@@ -10,13 +10,13 @@ import { UpdatePlayerJobPayload } from './instances/UpdatePlayerJob';
 export enum JobType {
   ASSERT_PLAYER_TYPE = 'ASSERT_PLAYER_TYPE',
   INVALIDATE_PERIOD_DELTAS = 'INVALIDATE_PERIOD_DELTAS',
-  REFRESH_NAME_CHANGES = 'REFRESH_NAME_CHANGES',
   REVIEW_NAME_CHANGE = 'REVIEW_NAME_CHANGE',
   REVIEW_PLAYER_TYPE = 'REVIEW_PLAYER_TYPE',
   SCHEDULE_COMPETITION_EVENTS = 'SCHEDULE_COMPETITION_EVENTS',
   SCHEDULE_COMPETITION_SCORE_UPDATES = 'SCHEDULE_COMPETITION_SCORE_UPDATES',
   SCHEDULE_DELTA_INVALIDATIONS = 'SCHEDULE_DELTA_INVALIDATIONS',
   SCHEDULE_GROUP_SCORE_UPDATES = 'SCHEDULE_GROUP_SCORE_UPDATES',
+  SCHEDULE_NAME_CHANGE_REVIEWS = 'SCHEDULE_NAME_CHANGE_REVIEWS',
   UPDATE_COMPETITION_SCORE = 'UPDATE_COMPETITION_SCORE',
   UPDATE_GROUP_SCORE = 'UPDATE_GROUP_SCORE',
   UPDATE_PLAYER = 'UPDATE_PLAYER'
